+++ madhan-lorry-owner/src/app/manage-driver/manage-driver-list/manage-driver-list.component.ts
@@ -1,8 +1,9 @@
 import { Component } from '@angular/core';
 import { Router } from '@angular/router';
 import { ManageDriverApiService } from '../services/api/manage-driver-api.service';
-import { ToastController } from '@ionic/angular';
 import { LocalStorageService } from 'src/app/services/local-storage/local-storage.service';
+import { ToastService } from 'src/app/services/toast/toast.service';
+import { AlertServiceService } from 'src/app/services/alert/alert-service.service';
 
 @Component({
   selector: 'app-manage-driver-list',
@@ -18,7 +19,8 @@ export class ManageDriverListComponent {
   constructor(private router: Router,
     private driverApi: ManageDriverApiService,
     private ls: LocalStorageService,
-    private toaster: ToastController) { }
+    private toast: ToastService,
+    private alert: AlertServiceService) { }
 
   ngOnInit() { }
 
@@ -46,26 +48,17 @@ export class ManageDriverListComponent {
     this.router.navigate(['manage-driver', 'edit', driver.driverId, driver.userId]);
   }
   deleteDriver(driverId) {
-    this.driverDeleteJson.driverId = driverId;
-    this.driverDeleteJson.refrefCreatedBy = this.userId;
-    this.driverApi.deleteDriver(this.driverDeleteJson, driverId).subscribe(success => {
-      console.log('success', success);
-      this.successToaster(success[0].msg);
-      this.ionViewWillEnter();
-    }, failure => { });
-  }
-  async successToaster(message) {
-    console.log('inside-->');
-    let toast = await this.toaster.create({
-      message: message,
-      duration: 2000,
-      position: 'top',
-      animated: true,
-      color: "success",
-      mode: "ios"
+    this.alert.alertPromt().then(data => {
+      if (Boolean(data)) {
+        this.driverDeleteJson.driverId = driverId;
+        this.driverDeleteJson.refrefCreatedBy = this.userId;
+        this.driverApi.deleteDriver(this.driverDeleteJson, driverId).subscribe(success => {
+          console.log('success', success);
+          this.toast.success(success[0].msg);
+          this.ionViewWillEnter();
+        }, failure => { });
+      }
     });
-
-
-    toast.present();
   }
+
 }
